+++ agents_playground/terminal/constants.py
@@ -1,0 +1,2 @@
+# The maximum number of arguments allowed in a function call.
+TERM_MAX_ARG_SIZE = 255
