+++ agents_playground/terminal/parser.py
@@ -8,6 +8,7 @@ from agents_playground.terminal.ast.statements import (
   Break,
   Continue,
   Expression,
+  Function,
   If,
   Stmt, 
   Clear, 
@@ -18,6 +19,7 @@ from agents_playground.terminal.ast.statements import (
 )
 from agents_playground.terminal.ast.expressions import ( 
   Assign,
+  Call,
   Expr,
   BinaryExpr, 
   GroupingExpr, 
@@ -26,6 +28,7 @@ from agents_playground.terminal.ast.expressions import (
   UnaryExpr,
   Variable
 )
+from agents_playground.terminal.constants import TERM_MAX_ARG_SIZE
 from agents_playground.terminal.token import Token
 from agents_playground.terminal.token_type import TokenType
 
@@ -62,17 +65,44 @@ class Parser:
 
   """
   Implements Grammar Rule
-  declaration -> varDecl | statement ;
+  declaration -> funDecl | varDecl | statement ;
   """
   def _declaration(self) -> Stmt | None:
     try:
-      if self._match(TokenType.VAR):
+      if self._match(TokenType.FUNC):
+        return self._function("function")
+      elif self._match(TokenType.VAR):
         return self._var_declaration()
       else:
         return self._statement()
     except ParseError as pe:
       self._synchronize()
       return None
+
+  def _function(self, kind: str) -> Function:
+    # 1. Consume the name of the function declaration.
+    name: Token = self._consume(TokenType.IDENTIFIER, f"Expect {kind} name.")
+
+    # 2. Consume the opening ( of the function declaration.
+    self._consume(TokenType.LEFT_PAREN, f"Expect '(' after {kind} name.")
+
+    # 3. Process all of the parameters.
+    parameters: List[Token] = []
+    if not self._check(TokenType.RIGHT_PAREN):
+      while True:
+        if len(parameters) >= TERM_MAX_ARG_SIZE:
+          self._error(self._peek(), f"Cannot have more than {TERM_MAX_ARG_SIZE} parameters.")
+        parameters.append(self._consume(TokenType.IDENTIFIER, "Expect parameter name."))
+        if not self._match(TokenType.COMMA):
+          break;
+
+    # 4. Consume the closing ).
+    self._consume(TokenType.RIGHT_PAREN, "Expect ')' after parameters.")
+
+    # 5. Parse the body of the function declaration.
+    self._consume(TokenType.LEFT_BRACE, f"Expect '{{' before {kind} body.")
+    body: List[Stmt] = self._block()
+    return Function(name, parameters, body)
 
   def _synchronize(self) -> None:
     """When an error has occurred parsing, jump forward until the next viable token is discovered."""
@@ -399,18 +429,48 @@ class Parser:
 
   """
   Implements Grammar Rule
-  unary -> ( "!" | "-" ) unary | primary ;
+  unary -> ( "!" | "-" ) unary | call ;
   """
   def _unary(self) -> Expr:
     if self._match(TokenType.BANG, TokenType.MINUS):
       operator: Token = self._previous()
       right: Expr = self._unary()
       return UnaryExpr(operator, right)
-    return self._primary()
+    return self._call()
 
   """
   Implements Grammar Rule
-  primary -> NUMBER | STRING | "true" | "false" | "nil" | "(" expression ")" | IDENTIFIER;
+  call -> primary ( "(" arguments? ")" )* ;
+  """
+  def _call(self) -> Expr:
+    expr: Expr = self._primary()
+    while True:
+      if self._match(TokenType.LEFT_PAREN):
+        expr = self._finish_call(expr)
+      else:
+        break;
+    return expr
+
+  def _finish_call(self, callee: Expr) -> Expr:
+    arguments: List[Expr] = []
+    if not self._check(TokenType.RIGHT_PAREN):
+      while True:
+        if len(arguments) >= TERM_MAX_ARG_SIZE:
+          self._error(self._peek(), f'Cannot have more than {TERM_MAX_ARG_SIZE} TERM_MAX_ARG_SIZE.')
+        arguments.append(self._expression())
+        if not self._match(TokenType.COMMA):
+          break
+    paren: Token = self._consume(TokenType.RIGHT_PAREN, "Expect ')' after arguments.")
+    return Call(callee, paren, arguments)
+
+  """
+  Implements Grammar Rule
+  primary ->  NUMBER | 
+              STRING | 
+              "true" | "false" | 
+              "None" | 
+              "(" expression ")" | 
+              IDENTIFIER;
   """
   def _primary(self) -> Expr:
     if self._match(TokenType.FALSE):
